# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Papertrail API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/curate/dedupe/keep": {
      "post": {
        "tags": [
          "curate"
        ],
        "summary": "Keep Both",
        "description": "Say these documents are different purchases, so the cluster stops coming back.",
        "operationId": "keep_both_api_curate_dedupe_keep_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeepIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DedupeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "curate"
        ],
        "summary": "Consider Again",
        "description": "Undo that, so the pair is offered again.",
        "operationId": "consider_again_api_curate_dedupe_keep_delete",
        "parameters": [
          {
            "name": "first",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "First"
            }
          },
          {
            "name": "second",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Second"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DedupeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/curate/dedupe/toss": {
      "post": {
        "tags": [
          "curate"
        ],
        "summary": "Toss Duplicate",
        "description": "Move one archived document's pages into ``tossed/``. The scans stay on disk, out of the archive.",
        "operationId": "toss_duplicate_api_curate_dedupe_toss_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TossIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TossOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/curate/dedupe/restore": {
      "post": {
        "tags": [
          "curate"
        ],
        "summary": "Restore Tossed",
        "description": "Undo a toss: put the pages back in the archive under the verdict they had.",
        "operationId": "restore_tossed_api_curate_dedupe_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DedupeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/curate/normalize": {
      "get": {
        "tags": [
          "curate"
        ],
        "summary": "Normalize Clusters",
        "description": "Names one engine thinks are the same shop. Embedding clustering asks Ollama for any new name,\nso it takes the single model slot and is refused while a job holds it.",
        "operationId": "normalize_clusters_api_curate_normalize_get",
        "parameters": [
          {
            "name": "engine",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Engine"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalizeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/curate/normalize/preview": {
      "post": {
        "tags": [
          "curate"
        ],
        "summary": "Preview Merge",
        "description": "Exactly what the merge would rewrite and re-file, before anything is written.",
        "operationId": "preview_merge_api_curate_normalize_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/curate/normalize/merge": {
      "post": {
        "tags": [
          "curate"
        ],
        "summary": "Merge Names",
        "description": "Rewrite every variant to the chosen name and re-file the documents it renames.",
        "operationId": "merge_names_api_curate_normalize_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/curate/normalize/distinct": {
      "post": {
        "tags": [
          "curate"
        ],
        "summary": "Confirm Distinct",
        "description": "Remember that these names are different shops, so the cluster stops coming back.",
        "operationId": "confirm_distinct_api_curate_normalize_distinct_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistinctIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistinctOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "curate"
        ],
        "summary": "Forget Distinct",
        "operationId": "forget_distinct_api_curate_normalize_distinct_delete",
        "parameters": [
          {
            "name": "first",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "First"
            }
          },
          {
            "name": "second",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Second"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistinctOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DedupeCluster": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/DedupeMember"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "time",
          "members"
        ],
        "title": "DedupeCluster",
        "description": "Documents with the same cost within a few minutes of each other."
      },
      "DedupeMember": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "cost": {
            "type": "number",
            "title": "Cost"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "path",
          "name",
          "date",
          "time",
          "cost",
          "currency",
          "pages"
        ],
        "title": "DedupeMember"
      },
      "DedupeOut": {
        "properties": {
          "archived": {
            "type": "integer",
            "title": "Archived"
          },
          "tossed": {
            "type": "integer",
            "title": "Tossed"
          },
          "clusters": {
            "items": {
              "$ref": "#/components/schemas/DedupeCluster"
            },
            "type": "array",
            "title": "Clusters"
          },
          "kept": {
            "items": {
              "$ref": "#/components/schemas/KeptPair"
            },
            "type": "array",
            "title": "Kept"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "archived",
          "tossed",
          "clusters",
          "kept"
        ],
        "title": "DedupeOut"
      },
      "DistinctIn": {
        "properties": {
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "names"
        ],
        "title": "DistinctIn"
      },
      "DistinctOut": {
        "properties": {
          "pairs": {
            "type": "integer",
            "title": "Pairs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pairs"
        ],
        "title": "DistinctOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeepIn": {
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Documents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "KeepIn"
      },
      "KeptPair": {
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Documents"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "documents",
          "names"
        ],
        "title": "KeptPair",
        "description": "Two documents you said are different purchases, and their names for the undo list."
      },
...
      "MergeIn": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target"
          },
          "variants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Variants"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target",
          "variants"
        ],
        "title": "MergeIn"
      },
      "MergeOut": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target"
          },
          "variants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Variants"
          },
          "documents": {
            "type": "integer",
            "title": "Documents"
          },
          "moves": {
            "items": {
              "$ref": "#/components/schemas/MoveOut"
            },
            "type": "array",
            "title": "Moves"
          },
          "decisions": {
            "type": "integer",
            "title": "Decisions"
          },
          "smart_matches": {
            "type": "integer",
            "title": "Smart Matches"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target",
          "variants",
          "documents",
          "moves",
          "decisions",
          "smart_matches",
          "error"
        ],
        "title": "MergeOut",
        "description": "What a merge would do (preview) or did. ``moves`` is every document that is re-filed."
      },
...
      "MoveOut": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "destination"
        ],
        "title": "MoveOut"
      },
      "NameCount": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "count"
        ],
        "title": "NameCount"
      },
      "NameGroupOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          },
          "counts": {
            "items": {
              "$ref": "#/components/schemas/NameCount"
            },
            "type": "array",
            "title": "Counts"
          },
          "canonical": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Canonical"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "names",
          "counts",
          "canonical"
        ],
        "title": "NameGroupOut",
        "description": "Names one engine thinks are the same shop, with how many documents carry each."
      },
...
      "NormalizeEngineOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "NormalizeEngineOut"
      },
      "NormalizeOut": {
        "properties": {
          "engines": {
            "items": {
              "$ref": "#/components/schemas/NormalizeEngineOut"
            },
            "type": "array",
            "title": "Engines"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "names": {
            "type": "integer",
            "title": "Names"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/NameGroupOut"
            },
            "type": "array",
            "title": "Groups"
          },
          "distinct_pairs": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Distinct Pairs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "engines",
          "engine",
          "threshold",
          "names",
          "groups",
          "distinct_pairs"
        ],
        "title": "NormalizeOut"
      },
...
      "RestoreIn": {
        "properties": {
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Paths"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "accepted",
              "marked",
              "tossed"
            ],
            "title": "Verdict"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paths",
          "verdict"
        ],
        "title": "RestoreIn"
      },
...
      "TossIn": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "path"
        ],
        "title": "TossIn"
      },
      "TossOut": {
        "properties": {
          "tossed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tossed"
          },
          "previous_verdict": {
            "type": "string",
            "enum": [
              "accepted",
              "marked",
              "tossed"
            ],
            "title": "Previous Verdict"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tossed",
          "previous_verdict",
          "name"
        ],
        "title": "TossOut",
        "description": "Where the pages went, and the verdict they had — everything an undo needs."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}